electron: Move state management to the backend. [Part 1] (#349)
* Add state management to the backend.
* Rename, reorganize classes to be more organized.
* Implement an ABC for Game.
* Small fixes.
